Move tests related to offset pagination to their own class
